commit Slowly add and subtract gears

+++ Inc/Includes/dac8831.h
@@ -33,7 +33,7 @@ static HAL_DMA_MuxSyncConfigTypeDef dmamux_syncParams;
 
 //dac8831 set value output
 void Wave_select(uint8_t n, uint16_t *_pBuf);
-void Dac8831_Set_Amp(uint16_t amp, uint16_t *_pBuf);
+void Dac8831_Set_Amp(float amp, uint16_t *_pBuf);
 void DAC8831_Set_Data(uint16_t data);
 void DAC8831_Set_Data_Dma(uint16_t *_pbufch1, uint32_t _sizech1,uint32_t _ulFreq);
 void Dac_level_CTL(uint8_t work_mode);
